# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel list car api demo",
        "contact": {
            "email": "gulyavcevvv@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Demo list car api Server"
        }
    ],
    "paths": {
        "/api": {},
        "/api/user/register": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Register user",
                "operationId": "acc68602f9da04236f7f5f087a27ebbb",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "User name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string",
                                            "example": "true"
                                        },
                                        "token": {
                                            "properties": {
                                                "token": {
                                                    "type": "string",
                                                    "example": "AbCvTmwO7yP0h9sRz5KRUplvRfW34T1GKUMryFffBy"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user/token": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get new token",
                "operationId": "418dcdff484ccca0fb42e057bcb80be0",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string",
                                            "example": "true"
                                        },
                                        "token": {
                                            "properties": {
                                                "token": {
                                                    "type": "string",
                                                    "example": "AbCvTmwO7yP0h9sRz5KRUplvRfW34T1GKUMryFffBy"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/car": {
            "get": {
                "tags": [
                    "Car"
                ],
                "summary": "View list car",
                "operationId": "4ef39d2263517f5584f70288fe365829",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "path",
                        "description": "limit. Default 5",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "int": {
                                "summary": "An int value.",
                                "value": "1"
                            }
                        }
                    },
                    {
                        "name": "offset",
                        "in": "path",
                        "description": "offset. Default 0",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "int": {
                                "summary": "An int value.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string",
                                            "example": "true"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Car"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Car"
                ],
                "summary": "Create car",
                "operationId": "9e322213a2b3788ddea57a47c09837f3",
                "parameters": [
                    {
                        "name": "model",
                        "in": "query",
                        "description": "Model name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string",
                                            "example": "true"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Car"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/car/my": {
            "get": {
                "tags": [
                    "Car"
                ],
                "summary": "View my car",
                "operationId": "e1f1d9ae36593a8c08f0d964457efa54",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string",
                                            "example": "true"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Car"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/car/{id}": {
            "get": {
                "tags": [
                    "Car"
                ],
                "summary": "Show car by id",
                "operationId": "b471339ac1d2c4628bf5b97ea1cc3d1b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id car",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "int": {
                                "summary": "An int value.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string",
                                            "example": "true"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Car"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Car"
                ],
                "summary": "Updates car",
                "operationId": "125b4c366e3a00128243368d418b5bba",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id car",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "int": {
                                "summary": "An int value.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string",
                                            "example": "true"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Car"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Car"
                ],
                "summary": "Delete car",
                "operationId": "082e4fcb212b90a462309876eebdfbb1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id car",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "int": {
                                "summary": "An int value.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string",
                                            "example": "true"
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Car": {
                "required": [
                    "model"
                ],
                "properties": {
                    "id": {
                        "type": "BigInt",
                        "example": "1"
                    },
                    "model": {
                        "type": "string",
                        "example": "BMW x5"
                    },
                    "user_id": {
                        "type": "BigInt",
                        "example": "1"
                    }
                },
                "type": "object"
            }
        },
        "responses": {
            "400": {
                "description": "error",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "string",
                                    "example": "false"
                                },
                                "message": {
                                    "type": "string",
                                    "example": "The user already has a car"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Car",
            "description": "API Endpoints of car"
        },
        {
            "name": "User",
            "description": "API Endpoints of user"
        }
    ]
}